+++ js_modules/dagster-ui/packages/ui-core/src/app/AppTopNav/useJobStateForNav.tsx
@@ -1,7 +1,7 @@
 import {useMemo} from 'react';
 
 import {isHiddenAssetGroupJob} from '../../asset-graph/Utils';
-import {useRepositoryOptions} from '../../workspace/WorkspaceContext/WorkspaceContext';
+import {useRepositoryOptions} from '../../workspace/WorkspaceContext/util';
 
 export type JobStateForNav = 'unknown' | 'has-jobs' | 'no-jobs';
