+++ js_modules/dagster-ui/packages/ui-core/src/assets/LaunchAssetChoosePartitionsDialog.tsx
@@ -76,7 +76,7 @@ import {assembleIntoSpans, stringForSpan} from '../partitions/SpanRepresentation
 import {DagsterTag} from '../runs/RunTag';
 import {testId} from '../testing/testId';
 import {ToggleableSection} from '../ui/ToggleableSection';
-import {useFeatureFlagForCodeLocation} from '../workspace/WorkspaceContext/WorkspaceContext';
+import {useFeatureFlagForCodeLocation} from '../workspace/WorkspaceContext/util';
 import {RepoAddress} from '../workspace/types';
 
 const MISSING_FAILED_STATUSES = [AssetPartitionStatus.MISSING, AssetPartitionStatus.FAILED];
