+++ js_modules/dagster-ui/packages/ui-core/src/instance/useDaemonStatus.tsx
@@ -6,7 +6,7 @@ import {InstanceWarningQuery, InstanceWarningQueryVariables} from './types/useDa
 import {gql, useQuery} from '../apollo-client';
 import {FIFTEEN_SECONDS, useQueryRefreshAtInterval} from '../app/QueryRefresh';
 import {InstigationStatus} from '../graphql/types';
-import {useRepositoryOptions} from '../workspace/WorkspaceContext/WorkspaceContext';
+import {useRepositoryOptions} from '../workspace/WorkspaceContext/util';
 
 export const useDaemonStatus = (skip = false): StatusAndMessage | null => {
   const {options} = useRepositoryOptions();
