+++ js_modules/dagster-ui/packages/ui-core/src/nav/LeftNavItem.tsx
@@ -13,7 +13,7 @@ import {InstigationStatus} from '../graphql/types';
 import {INSTIGATION_STATE_BASE_FRAGMENT} from '../instigation/InstigationStateBaseFragment';
 import {InstigationStateFragment} from '../instigation/types/InstigationUtils.types';
 import {humanCronString} from '../schedules/humanCronString';
-import {useRepository} from '../workspace/WorkspaceContext/WorkspaceContext';
+import {useRepository} from '../workspace/WorkspaceContext/util';
 import {workspacePathFromAddress} from '../workspace/workspacePath';
 
 interface LeftNavItemProps {
