+++ js_modules/dagster-ui/packages/ui-core/src/overview/visibleRepoKeys.tsx
@@ -1,4 +1,4 @@
-import {DagsterRepoOption} from '../workspace/WorkspaceContext/WorkspaceContext';
+import {DagsterRepoOption} from '../workspace/WorkspaceContext/util';
 import {buildRepoAddress} from '../workspace/buildRepoAddress';
 import {repoAddressAsHumanString} from '../workspace/repoAddressAsString';
