+++ js_modules/dagster-ui/packages/ui-core/src/pipelines/PipelineExplorerRoot.oss.tsx
@@ -25,7 +25,7 @@ import {assetDetailsPathForKey} from '../assets/assetDetailsPathForKey';
 import {useDocumentTitle} from '../hooks/useDocumentTitle';
 import {METADATA_ENTRY_FRAGMENT} from '../metadata/MetadataEntryFragment';
 import {Loading} from '../ui/Loading';
-import {buildPipelineSelector} from '../workspace/WorkspaceContext/WorkspaceContext';
+import {buildPipelineSelector} from '../workspace/WorkspaceContext/util';
 import {RepoAddress} from '../workspace/types';
 
 export const PipelineExplorerSnapshotRoot = () => {
