+++ js_modules/dagster-ui/packages/ui-core/src/pipelines/SidebarContainerOverview.tsx
@@ -11,10 +11,7 @@ import {gql} from '../apollo-client';
 import {breakOnUnderscores} from '../app/Util';
 import {MetadataEntry} from '../metadata/MetadataEntry';
 import {METADATA_ENTRY_FRAGMENT} from '../metadata/MetadataEntryFragment';
-import {
-  findRepositoryAmongOptions,
-  useRepositoryOptions,
-} from '../workspace/WorkspaceContext/WorkspaceContext';
+import {findRepositoryAmongOptions, useRepositoryOptions} from '../workspace/WorkspaceContext/util';
 import {repoContainsPipeline} from '../workspace/findRepoContainingPipeline';
 import {RepoAddress} from '../workspace/types';
