+++ js_modules/dagster-ui/packages/ui-core/src/runs/RunsFilterInput.tsx
@@ -31,7 +31,7 @@ import {
   useSuggestionFilter,
 } from '../ui/BaseFilters/useSuggestionFilter';
 import {TimeRangeState, useTimeRangeFilter} from '../ui/BaseFilters/useTimeRangeFilter';
-import {useRepositoryOptions} from '../workspace/WorkspaceContext/WorkspaceContext';
+import {useRepositoryOptions} from '../workspace/WorkspaceContext/util';
 
 export interface RunsFilterInputProps {
   loading?: boolean;
