+++ js_modules/dagster-ui/packages/ui-core/src/schedules/SchedulePartitionStatus.tsx
@@ -14,7 +14,7 @@ import {gql, useLazyQuery} from '../apollo-client';
 import {assertUnreachable} from '../app/Util';
 import {RunStatus} from '../graphql/types';
 import {StatusTable} from '../instigation/InstigationUtils';
-import {isThisThingAJob, useRepository} from '../workspace/WorkspaceContext/WorkspaceContext';
+import {isThisThingAJob, useRepository} from '../workspace/WorkspaceContext/util';
 import {RepoAddress} from '../workspace/types';
 import {workspacePathFromAddress} from '../workspace/workspacePath';
