+++ js_modules/dagster-ui/packages/ui-core/src/schedules/SchedulesNextTicks.tsx
@@ -46,7 +46,7 @@ import {
   isThisThingAJob,
   useRepository,
   useRepositoryOptions,
-} from '../workspace/WorkspaceContext/WorkspaceContext';
+} from '../workspace/WorkspaceContext/util';
 import {repoAddressToSelector} from '../workspace/repoAddressToSelector';
 import {RepoAddress} from '../workspace/types';
 import {workspacePathFromAddress} from '../workspace/workspacePath';
