+++ js_modules/dagster-ui/packages/ui-core/src/typeexplorer/TypeExplorerContainer.tsx
@@ -6,7 +6,7 @@ import {
 import {gql, useQuery} from '../apollo-client';
 import {ExplorerPath} from '../pipelines/PipelinePathUtils';
 import {Loading} from '../ui/Loading';
-import {buildPipelineSelector} from '../workspace/WorkspaceContext/WorkspaceContext';
+import {buildPipelineSelector} from '../workspace/WorkspaceContext/util';
 import {RepoAddress} from '../workspace/types';
 
 interface ITypeExplorerContainerProps {
