+++ js_modules/dagster-ui/packages/ui-core/src/workspace/__tests__/getFeatureFlagForCodeLocation.test.tsx
@@ -1,6 +1,6 @@
 import {buildFeatureFlag, buildWorkspaceLocationEntry} from '../../graphql/types';
-import {getFeatureFlagForCodeLocation} from '../WorkspaceContext/WorkspaceContext';
 import {WorkspaceLocationNodeFragment} from '../WorkspaceContext/types/WorkspaceQueries.types';
+import {getFeatureFlagForCodeLocation} from '../WorkspaceContext/util';
 
 describe('getFeatureFlagForCodeLocation', () => {
   const locationEntries: WorkspaceLocationNodeFragment[] = [
