+++ js_modules/dagster-ui/packages/ui-core/src/workspace/useRepositoryForRun.ts
@@ -1,4 +1,4 @@
-import {DagsterRepoOption, useRepositoryOptions} from './WorkspaceContext/WorkspaceContext';
+import {DagsterRepoOption, useRepositoryOptions} from './WorkspaceContext/util';
 import {findRepoContainingPipeline, repoContainsPipeline} from './findRepoContainingPipeline';
 
 type MatchType = {
